requirements.txt
altgraph==0.17.4
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.1
idna==3.10
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
keyring==25.6.0
more-itertools==10.8.0
numpy==2.3.3
oracledb==3.3.0
packaging==25.0
pandas==2.3.2
pefile==2023.2.7
pipreqs==0.4.13
pycparser==2.23
pycryptodome==3.23.0
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.8
PyQt6==6.9.1
PyQt6-Qt6==6.9.2
PyQt6_sip==13.10.2
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32-ctypes==0.2.3
requests==2.32.5
setuptools==80.9.0
six==1.17.0
sql-formatter==0.6.2
tzdata==2025.2
urllib3==2.5.0
yarg==0.1.10
